% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mht.R
\name{plot_mht}
\alias{plot_mht}
\title{Zoom in manhattan plot}
\usage{
plot_mht(grObj, wh)
}
\arguments{
\item{grObj}{GRanges object with scores as the -log10 p-value}

\item{wh}{genome interval to plot}
}
\description{
Zoom in manhattan plot
}
