% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBed.R
\name{plotBed}
\alias{plotBed}
\title{Plot BED intervals}
\usage{
plotBed(grObj, wh)
}
\arguments{
\item{grObj}{GRanges object with scores as the posterior inclusion probability}

\item{wh}{genome interval to plot}
}
\description{
Plot BED intervals
}
