% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plotEnsGenes}
\alias{plotEnsGenes}
\title{Plot ENSEMBL genes}
\usage{
plotEnsGenes(
  ensdb,
  wh,
  splice_variants = FALSE,
  non_coding = FALSE,
  arrow.size = 0.05,
  ensGene = NULL
)
}
\arguments{
\item{ensdb}{ENSEMBL database object like EnsDb.Hsapiens.v86}

\item{wh}{genome interval to plot}

\item{splice_variants}{if TRUE, show multiple transcripts from the same gene}

\item{non_coding}{if TRUE, also show non-coding genes}

\item{arrow.size}{arrow size for genes}

\item{ensGene}{show specific non-coding locus}
}
\value{
ggplot2 of genome region
}
\description{
Plot ENSEMBL genes in region
}
\examples{
library(EnsDb.Hsapiens.v86)
library(GenomicRanges) 

# gene database
ensdb = EnsDb.Hsapiens.v86

# interval
query = GRanges("20", IRanges(62045027,62164563))

# plot genes
plotEnsGenes( ensdb, query)

}
