% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMht.R
\name{plotMht}
\alias{plotMht}
\title{Zoom in manhattan plot}
\usage{
plotMht(grObj, wh, size = 8)
}
\arguments{
\item{grObj}{GRanges object with scores as the -log10 p-value}

\item{wh}{genome interval to plot}

\item{size}{font size}
}
\description{
Zoom in manhattan plot
}
