% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosterior.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Zoom in manhattan plot of posteriors}
\usage{
plotPosterior(grObj, wh, size = 8)
}
\arguments{
\item{grObj}{GRanges object with scores as the posterior inclusion probability}

\item{wh}{genome interval to plot}

\item{size}{font size}
}
\description{
Zoom in manhattan plot of posteriors
}
